Fix module collision by renaming loaders.py to loader_api.py
