+++ pyproject.toml
@@ -147,7 +147,7 @@ warn_redundant_casts = true
 warn_unused_ignores = false
 warn_no_return = true
 warn_unreachable = false
-exclude = ["scripts", "src/llm_rag/document_processing/loaders/__init__.py"]  # Exclude the duplicate module
+exclude = ["scripts"]  # Exclude scripts directory
 namespace_packages = true
 explicit_package_bases = true
 error_summary = false
